requirements.txt
# Core backend
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.0

# Data & feature engineering
numpy>=1.26.0
pandas>=2.2.0

# Machine Learning 
scikit-learn>=1.4.0
xgboost>=2.0.0
joblib>=1.3.0

# Explainability
shap>=0.45.0

# Data ingestion
kagglehub>=0.2.6

# Validation & schemas
pydantic>=2.7.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Database
psycopg2-binary

# LLM Integration
openai